requirements.txt
streamlit-0.84.0-py2.py3-none-any.whl
matplotlib